added new lookup modules for user, org, vc + account to try and isolate out module dependencies
