+++ src/domain/access/invitation/invitation.module.ts
@@ -3,25 +3,25 @@ import { Invitation } from '@domain/access/invitation';
 import { InvitationService } from '@domain/access/invitation/invitation.service';
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UserModule } from '@domain/community/user/user.module';
 import { AuthorizationPolicyModule } from '@domain/common/authorization-policy/authorization.policy.module';
 import { InvitationResolverFields } from './invitation.resolver.fields';
 import { AuthorizationModule } from '@core/authorization/authorization.module';
 import { InvitationAuthorizationService } from './invitation.service.authorization';
 import { InvitationResolverMutations } from './invitation.resolver.mutations';
 import { ContributorModule } from '@domain/community/contributor/contributor.module';
-import { VirtualContributorModule } from '@domain/community/virtual-contributor/virtual.contributor.module';
 import { InvitationLifecycleResolverFields } from './invitation.resolver.fields.lifecycle';
 import { InvitationLifecycleService } from './invitation.service.lifecycle';
+import { AccountLookupModule } from '@domain/space/account.lookup/account.lookup.module';
+import { VirtualContributorLookupModule } from '@domain/community/virtual-contributor-lookup/virtual.contributor.lookup.module';
 
 @Module({
   imports: [
     AuthorizationPolicyModule,
     AuthorizationModule,
     LifecycleModule,
-    UserModule,
     ContributorModule,
-    VirtualContributorModule,
+    VirtualContributorLookupModule,
+    AccountLookupModule,
     TypeOrmModule.forFeature([Invitation]),
   ],
   providers: [
