+++ src/domain/access/invitation/invitation.service.authorization.ts
@@ -5,11 +5,12 @@ import { AuthorizationPolicyService } from '@domain/common/authorization-policy/
 import { IAuthorizationPolicy } from '@domain/common/authorization-policy/authorization.policy.interface';
 import { IInvitation } from './invitation.interface';
 import { IAuthorizationPolicyRuleCredential } from '@core/authorization/authorization.policy.rule.credential.interface';
-import { CREDENTIAL_RULE_COMMUNITY_USER_INVITATION as CREDENTIAL_RULE_COMMUNITY_CONTRIBUTOR_INVITATION } from '@common/constants/authorization/credential.rule.constants';
 import { CommunityContributorType } from '@common/enums/community.contributor.type';
 import { ICredentialDefinition } from '@domain/agent/credential/credential.definition.interface';
-import { VirtualContributorService } from '@domain/community/virtual-contributor/virtual.contributor.service';
 import { ContributorService } from '@domain/community/contributor/contributor.service';
+import { AccountLookupService } from '@domain/space/account.lookup/account.lookup.service';
+import { VirtualContributorLookupService } from '@domain/community/virtual-contributor-lookup/virtual.contributor.lookup.service';
+import { CREDENTIAL_RULE_COMMUNITY_USER_INVITATION } from '@common/constants';
 
 @Injectable()
 export class InvitationAuthorizationService {
@@ -17,7 +18,8 @@ export class InvitationAuthorizationService {
     private invitationService: InvitationService,
     private authorizationPolicyService: AuthorizationPolicyService,
     private contributorService: ContributorService,
-    private virtualContributorService: VirtualContributorService
+    private virtualContributorLookupService: VirtualContributorLookupService,
+    private accountLookupService: AccountLookupService
   ) {}
 
   async applyAuthorizationPolicy(
@@ -66,10 +68,13 @@ export class InvitationAuthorizationService {
         });
         break;
       case CommunityContributorType.VIRTUAL:
-        const vcHostCriterias =
-          await this.virtualContributorService.getAccountHostCredentials(
+        const account =
+          await this.virtualContributorLookupService.getAccountOrFail(
             contributor.id
           );
+        const vcHostCriterias =
+          await this.accountLookupService.getHostCredentials(account);
+
         criterias.push(...vcHostCriterias);
         break;
     }
@@ -82,7 +87,7 @@ export class InvitationAuthorizationService {
           AuthorizationPrivilege.COMMUNITY_INVITE_ACCEPT,
         ],
         criterias,
-        CREDENTIAL_RULE_COMMUNITY_CONTRIBUTOR_INVITATION
+        CREDENTIAL_RULE_COMMUNITY_USER_INVITATION
       );
     newRules.push(virtualInvitationRule);
