+++ src/domain/communication/message.reaction/message.reaction.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { MessageReactionResolverFields } from './message.reaction.resolver.fields';
-import { ContributorLookupModule } from '@services/infrastructure/contributor-lookup/contributor.lookup.module';
+import { UserLookupModule } from '@domain/community/user-lookup/user.lookup.module';
 
 @Module({
-  imports: [ContributorLookupModule],
+  imports: [UserLookupModule],
   providers: [MessageReactionResolverFields],
   exports: [MessageReactionResolverFields],
 })
