+++ src/domain/communication/message/message.module.ts
@@ -1,14 +1,10 @@
 import { Module } from '@nestjs/common';
 import { MessageResolverFields } from './message.resolver.fields';
-import { ContributorLookupModule } from '@services/infrastructure/contributor-lookup/contributor.lookup.module';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { VirtualContributor } from '@domain/community/virtual-contributor/virtual.contributor.entity';
+import { UserLookupModule } from '@domain/community/user-lookup/user.lookup.module';
+import { VirtualContributorLookupModule } from '@domain/community/virtual-contributor-lookup/virtual.contributor.lookup.module';
 
 @Module({
-  imports: [
-    ContributorLookupModule,
-    TypeOrmModule.forFeature([VirtualContributor]),
-  ],
+  imports: [UserLookupModule, VirtualContributorLookupModule],
   providers: [MessageResolverFields],
   exports: [MessageResolverFields],
 })
