+++ src/domain/community/organization-lookup/organization.lookup.service.ts
@@ -1,0 +1,67 @@
+import { EntityManager, FindOneOptions } from 'typeorm';
+import { InjectEntityManager } from '@nestjs/typeorm';
+import { Inject, LoggerService } from '@nestjs/common';
+import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
+import { EntityNotFoundException } from '@common/exceptions';
+import { LogContext } from '@common/enums';
+import { Organization } from '../organization/organization.entity';
+import { IOrganization } from '../organization/organization.interface';
+
+export class OrganizationLookupService {
+  constructor(
+    @InjectEntityManager('default')
+    private entityManager: EntityManager,
+    @Inject(WINSTON_MODULE_NEST_PROVIDER)
+    private readonly logger: LoggerService
+  ) {}
+
+  async getOrganizationByUUID(
+    organizationID: string,
+    options?: FindOneOptions<Organization>
+  ): Promise<IOrganization | null> {
+    const organization: IOrganization | null = await this.entityManager.findOne(
+      Organization,
+      {
+        ...options,
+        where: { ...options?.where, id: organizationID },
+      }
+    );
+
+    return organization;
+  }
+
+  async getOrganizationByNameIdOrFail(
+    organizationNameID: string,
+    options?: FindOneOptions<Organization>
+  ): Promise<IOrganization> {
+    const organization: IOrganization | null = await this.entityManager.findOne(
+      Organization,
+      {
+        ...options,
+        where: { ...options?.where, nameID: organizationNameID },
+      }
+    );
+    if (!organization)
+      throw new EntityNotFoundException(
+        `Unable to find Organization with NameID: ${organizationNameID}`,
+        LogContext.COMMUNITY
+      );
+    return organization;
+  }
+
+  async getOrganizationOrFail(
+    organizationID: string,
+    options?: FindOneOptions<Organization>
+  ): Promise<IOrganization | never> {
+    const organization = await this.getOrganizationByUUID(
+      organizationID,
+      options
+    );
+    if (!organization)
+      throw new EntityNotFoundException(
+        `Unable to find Organization with ID: ${organizationID}`,
+        LogContext.COMMUNITY
+      );
+    return organization;
+  }
+}
