+++ src/domain/community/organization-role/organization.role.module.ts
@@ -6,15 +6,15 @@ import { UserModule } from '@domain/community/user/user.module';
 import { AuthorizationModule } from '@core/authorization/authorization.module';
 import { AgentModule } from '@domain/agent/agent/agent.module';
 import { AuthorizationPolicyModule } from '@domain/common/authorization-policy/authorization.policy.module';
-import { ContributorLookupModule } from '@services/infrastructure/contributor-lookup/contributor.lookup.module';
 import { OrganizationRoleAuthorizationService } from './organization.role.service.authorization';
+import { OrganizationLookupModule } from '../organization-lookup/organization.lookup.module';
 
 @Module({
   imports: [
     AgentModule,
     AuthorizationPolicyModule,
     AuthorizationModule,
-    ContributorLookupModule,
+    OrganizationLookupModule,
     UserModule,
   ],
   providers: [
