+++ src/domain/community/user-lookup/user.lookup.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { UserLookupService } from './user.lookup.service';
+
+@Module({
+  imports: [], // Important this is empty!
+  providers: [UserLookupService],
+  exports: [UserLookupService],
+})
+export class UserLookupModule {}
