+++ src/domain/community/virtual-contributor/virtual.contributor.service.ts
@@ -28,28 +28,22 @@ import { CommunicationAdapter } from '@services/adapters/communication-adapter/c
 import { NamingService } from '@services/infrastructure/naming/naming.service';
 import { AiPersonaService } from '../ai-persona/ai.persona.service';
 import { CreateAiPersonaInput } from '../ai-persona/dto';
-import {
-  InvocationResultAction,
-  VirtualContributorInvocationInput,
-  isInputValidForAction,
-} from './dto';
 import { AgentInfo } from '@core/authentication.agent.info/agent.info';
 import { AiServerAdapter } from '@services/adapters/ai-server-adapter/ai.server.adapter';
-import { AiServerAdapterInvocationInput } from '@services/adapters/ai-server-adapter/dto/ai.server.adapter.dto.invocation';
 import { SearchVisibility } from '@common/enums/search.visibility';
 import { IAiPersona } from '../ai-persona';
 import { IContributor } from '../contributor/contributor.interface';
 import { AccountHostService } from '@domain/space/account.host/account.host.service';
 import { ICredentialDefinition } from '@domain/agent/credential/credential.definition.interface';
-
 import { AgentType } from '@common/enums/agent.type';
 import { ContributorService } from '../contributor/contributor.service';
 import { AuthorizationPolicyType } from '@common/enums/authorization.policy.type';
 import { Invitation } from '@domain/access/invitation/invitation.entity';
 import { IStorageBucket } from '@domain/storage/storage-bucket/storage.bucket.interface';
-import { VcInteractionService } from '@domain/communication/vc-interaction/vc.interaction.service';
 import { IKnowledgeBase } from '@domain/common/knowledge-base/knowledge.base.interface';
 import { KnowledgeBaseService } from '@domain/common/knowledge-base/knowledge.base.service';
+import { AccountLookupService } from '@domain/space/account.lookup/account.lookup.service';
+import { VirtualContributorLookupService } from '../virtual-contributor-lookup/virtual.contributor.lookup.service';
 
 @Injectable()
 export class VirtualContributorService {
@@ -63,8 +57,9 @@ export class VirtualContributorService {
     private aiPersonaService: AiPersonaService,
     private aiServerAdapter: AiServerAdapter,
     private accountHostService: AccountHostService,
-    private vcInteractionService: VcInteractionService,
     private knowledgeBaseService: KnowledgeBaseService,
+    private virtualContributorLookupService: VirtualContributorLookupService,
+    private accountLookupService: AccountLookupService,
     @InjectEntityManager('default')
     private entityManager: EntityManager,
     @InjectRepository(VirtualContributor)
@@ -392,61 +387,6 @@ export class VirtualContributorService {
     );
   }
 
-  public async invoke(
-    invocationInput: VirtualContributorInvocationInput
-  ): Promise<void> {
-    const virtualContributor = await this.getVirtualContributorOrFail(
-      invocationInput.virtualContributorID,
-      {
-        relations: {
-          authorization: true,
-          aiPersona: true,
-          agent: true,
-          profile: true,
-        },
-      }
-    );
-    if (!virtualContributor.agent) {
-      throw new EntityNotInitializedException(
-        `Virtual Contributor Agent not initialized: ${invocationInput.virtualContributorID}`,
-        LogContext.AUTH
-      );
-    }
-
-    this.logger.verbose?.(
-      `still need to use the context ${invocationInput.contextSpaceID}, ${invocationInput.userID}`,
-      LogContext.AI_PERSONA_SERVICE_ENGINE
-    );
-
-    const aiServerAdapterInvocationInput: AiServerAdapterInvocationInput = {
-      aiPersonaServiceID: virtualContributor.aiPersona.aiPersonaServiceID,
-      message: invocationInput.message,
-      contextID: invocationInput.contextSpaceID,
-      userID: invocationInput.userID,
-      description: virtualContributor.profile.description,
-      displayName: virtualContributor.profile.displayName,
-      resultHandler: invocationInput.resultHandler,
-    };
-
-    if (
-      isInputValidForAction(invocationInput, InvocationResultAction.POST_REPLY)
-    ) {
-      const vcInteraction =
-        await this.vcInteractionService.getVcInteractionOrFail(
-          invocationInput.resultHandler.roomDetails!.vcInteractionID!
-        );
-
-      aiServerAdapterInvocationInput.vcInteractionID = vcInteraction.id;
-      aiServerAdapterInvocationInput.externalMetadata =
-        vcInteraction.externalMetadata;
-    }
-
-    const response = await this.aiServerAdapter.invoke(
-      aiServerAdapterInvocationInput
-    );
-    return response;
-  }
-
   // TODO: move to store
   async getVirtualContributors(
     args: ContributorQueryArgs
@@ -505,37 +445,24 @@ export class VirtualContributorService {
   public async getProvider(
     virtualContributor: IVirtualContributor
   ): Promise<IContributor> {
-    const virtualContributorWithAccount =
-      await this.getVirtualContributorOrFail(virtualContributor.id, {
-        relations: { account: true },
-      });
-    const account = virtualContributorWithAccount.account;
-    if (!account)
-      throw new EntityNotInitializedException(
-        `Virtual Contributor Account not initialized: ${virtualContributor.id}`,
-        LogContext.AUTH
-      );
+    const account = await this.virtualContributorLookupService.getAccountOrFail(
+      virtualContributor.id
+    );
 
-    const host = await this.accountHostService.getHostOrFail(account);
+    const host = await this.accountLookupService.getHostOrFail(account);
     return host;
   }
 
   public async getAccountHostCredentials(
     virtualContributorID: string
   ): Promise<ICredentialDefinition[]> {
-    const virtualContributorWithAccount =
-      await this.getVirtualContributorOrFail(virtualContributorID, {
-        relations: { account: true },
-      });
-    const account = virtualContributorWithAccount.account;
-    if (!account)
-      throw new EntityNotInitializedException(
-        `Virtual Contributor Account not initialized: ${virtualContributorID}`,
-        LogContext.AUTH
+    const account =
+      await this.virtualContributorLookupService.getAccountOrFail(
+        virtualContributorID
       );
 
     const hostCredentials =
-      await this.accountHostService.getHostCredentials(account);
+      await this.accountLookupService.getHostCredentials(account);
     return hostCredentials;
   }
 
@@ -585,33 +512,6 @@ export class VirtualContributorService {
     }
 
     return aiPersona;
-  }
-
-  async virtualContributorsWithCredentials(
-    credentialCriteria: CredentialsSearchInput
-  ): Promise<IVirtualContributor[]> {
-    const credResourceID = credentialCriteria.resourceID || '';
-    const virtualContributorMatches = await this.virtualContributorRepository
-      .createQueryBuilder('virtual_contributor')
-      .leftJoinAndSelect('virtual_contributor.agent', 'agent')
-      .leftJoinAndSelect('agent.credentials', 'credential')
-      .where('credential.type = :type')
-      .andWhere('credential.resourceID = :resourceID')
-      .setParameters({
-        type: `${credentialCriteria.type}`,
-        resourceID: credResourceID,
-      })
-      .getMany();
-
-    // reload to go through the normal loading path
-    const results: IVirtualContributor[] = [];
-    for (const virtualContributor of virtualContributorMatches) {
-      const loadedVirtual = await this.getVirtualContributorOrFail(
-        virtualContributor.id
-      );
-      results.push(loadedVirtual);
-    }
-    return results;
   }
 
   public async createVirtualContributorNameID(
