+++ src/domain/space/space/space.service.spec.ts
@@ -23,6 +23,7 @@ import { getRepositoryToken } from '@nestjs/typeorm';
 import { ISpaceSettings } from '@domain/space/space.settings/space.settings.interface';
 import { SpacePrivacyMode } from '@common/enums/space.privacy.mode';
 import { CommunityMembershipPolicy } from '@common/enums/community.membership.policy';
+import { CalloutsSet } from '@domain/collaboration/callouts-set/callouts.set.entity';
 
 const moduleMocker = new ModuleMocker(global);
 
@@ -298,24 +299,29 @@ const getSubspacesMock = (
       collaboration: {
         id: '',
         isTemplate: false,
-        groupsStr: JSON.stringify([
-          {
-            displayName: 'HOME',
-            description: 'The Home page.',
-          },
-          {
-            displayName: 'COMMUNITY',
-            description: 'The Community page.',
-          },
-          {
-            displayName: 'SUBSPACES',
-            description: 'The Subspaces page.',
-          },
-          {
-            displayName: 'KNOWLEDGE',
-            description: 'The knowledge page.',
-          },
-        ]),
+        calloutsSet: {
+          id: '',
+          callouts: [],
+          groupsStr: JSON.stringify([
+            {
+              displayName: 'HOME',
+              description: 'The Home page.',
+            },
+            {
+              displayName: 'COMMUNITY',
+              description: 'The Community page.',
+            },
+            {
+              displayName: 'SUBSPACES',
+              description: 'The Subspaces page.',
+            },
+            {
+              displayName: 'KNOWLEDGE',
+              description: 'The knowledge page.',
+            },
+          ]),
+          ...getEntityMock<CalloutsSet>(),
+        },
         innovationFlow: {
           id: '',
           states: JSON.stringify([
@@ -395,24 +401,29 @@ const getSubsubspacesMock = (subsubspaceId: string, count: number): Space[] => {
       collaboration: {
         id: '',
         isTemplate: false,
-        groupsStr: JSON.stringify([
-          {
-            displayName: 'HOME',
-            description: 'The Home page.',
-          },
-          {
-            displayName: 'COMMUNITY',
-            description: 'The Community page.',
-          },
-          {
-            displayName: 'SUBSPACES',
-            description: 'The Subspaces page.',
-          },
-          {
-            displayName: 'KNOWLEDGE',
-            description: 'The knowledge page.',
-          },
-        ]),
+        calloutsSet: {
+          id: '',
+          callouts: [],
+          groupsStr: JSON.stringify([
+            {
+              displayName: 'HOME',
+              description: 'The Home page.',
+            },
+            {
+              displayName: 'COMMUNITY',
+              description: 'The Community page.',
+            },
+            {
+              displayName: 'SUBSPACES',
+              description: 'The Subspaces page.',
+            },
+            {
+              displayName: 'KNOWLEDGE',
+              description: 'The knowledge page.',
+            },
+          ]),
+          ...getEntityMock<CalloutsSet>(),
+        },
         innovationFlow: {
           id: '',
           states: JSON.stringify([
