+++ src/domain/space/space/space.service.ts
@@ -59,10 +59,8 @@ import { CommunityContributorType } from '@common/enums/community.contributor.ty
 import { IStorageAggregator } from '@domain/storage/storage-aggregator/storage.aggregator.interface';
 import { AgentType } from '@common/enums/agent.type';
 import { StorageAggregatorType } from '@common/enums/storage.aggregator.type';
-import { AccountHostService } from '../account.host/account.host.service';
 import { AuthorizationPolicyType } from '@common/enums/authorization.policy.type';
 import { LicensingCredentialBasedCredentialType } from '@common/enums/licensing.credential.based.credential.type';
-import { LicensingCredentialBasedService } from '@platform/licensing/credential-based/licensing-credential-based-entitlements-engine/licensing.credential.based.service';
 import { ISpaceSubscription } from './space.license.subscription.interface';
 import { IAccount } from '../account/account.interface';
 import { LicensingCredentialBasedPlanType } from '@common/enums/licensing.credential.based.plan.type';
@@ -85,6 +83,7 @@ import { getDiff, hasOnlyAllowedFields } from '@common/utils';
 import { ILicensePlan } from '@platform/licensing/credential-based/license-plan/license.plan.interface';
 import { SpacePrivacyMode } from '@common/enums/space.privacy.mode';
 import { ICalloutsSet } from '@domain/collaboration/callouts-set/callouts.set.interface';
+import { AccountLookupService } from '../account.lookup/account.lookup.service';
 
 const EXPLORE_SPACES_LIMIT = 30;
 const EXPLORE_SPACES_ACTIVITY_DAYS_OLD = 30;
@@ -99,7 +98,7 @@ type SpaceSortingData = {
 @Injectable()
 export class SpaceService {
   constructor(
-    private accountHostService: AccountHostService,
+    private accountLookupService: AccountLookupService,
     private authorizationPolicyService: AuthorizationPolicyService,
     private spacesFilterService: SpaceFilterService,
     private contextService: ContextService,
@@ -114,7 +113,6 @@ export class SpaceService {
     private templatesManagerService: TemplatesManagerService,
     private collaborationService: CollaborationService,
     private licensingFrameworkService: LicensingFrameworkService,
-    private licenseEngineService: LicensingCredentialBasedService,
     private licenseService: LicenseService,
     @InjectRepository(Space)
     private spaceRepository: Repository<Space>,
@@ -1344,7 +1342,7 @@ export class SpaceService {
         LogContext.LIBRARY
       );
     }
-    const provider = await this.accountHostService.getHost(space.account);
+    const provider = await this.accountLookupService.getHost(space.account);
     if (!provider) {
       throw new RelationshipNotFoundException(
         `Unable to load provider for Space ${space.id} `,
