+++ src/platform/admin/licensing/admin.licensing.service.ts
@@ -16,11 +16,13 @@ import { ValidationException } from '@common/exceptions';
 import { LicensingFrameworkService } from '@platform/licensing/credential-based/licensing-framework/licensing.framework.service';
 import { EntityManager } from 'typeorm';
 import { InjectEntityManager } from '@nestjs/typeorm';
+import { AccountLookupService } from '@domain/space/account.lookup/account.lookup.service';
 
 @Injectable()
 export class AdminLicensingService {
   constructor(
     private accountHostService: AccountHostService,
+    private accountLookupService: AccountLookupService,
     private licensingFrameworkService: LicensingFrameworkService,
     private licenseIssuerService: LicenseIssuerService,
     private spaceService: SpaceService,
@@ -137,7 +139,7 @@ export class AdminLicensingService {
       );
     }
     // Todo: assign the actual credential for the license plan
-    const account = await this.accountHostService.getAccountOrFail(
+    const account = await this.accountLookupService.getAccountOrFail(
       licensePlanData.accountID,
       {
         relations: {
@@ -181,7 +183,7 @@ export class AdminLicensingService {
     }
 
     // Todo: assign the actual credential for the license plan
-    const account = await this.accountHostService.getAccountOrFail(
+    const account = await this.accountLookupService.getAccountOrFail(
       licensePlanData.accountID,
       {
         relations: {
