+++ src/services/infrastructure/contributor-lookup/contributor.lookup.module.ts
@@ -1,8 +1,9 @@
 import { Module } from '@nestjs/common';
 import { ContributorLookupService } from './contributor.lookup.service';
+import { UserLookupModule } from '@domain/community/user-lookup/user.lookup.module';
 
 @Module({
-  imports: [],
+  imports: [UserLookupModule],
   providers: [ContributorLookupService],
   exports: [ContributorLookupService],
 })
